requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI
langchain==0.0.350
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.0

# Data
pymongo==4.6.0
redis==5.0.1
faiss-cpu==1.7.4
motor==3.3.2

# Git & Parsing
GitPython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Notifications
slack-sdk==3.26.1
python-dotenv==1.0.0
httpx==0.25.2

# Observability
prometheus-client==0.19.0
langsmith==0.0.70

# Testing / Dev
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
=======
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI
langchain==0.0.350
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.0

# Data
pymongo==4.6.0
redis==5.0.1
faiss-cpu==1.7.4
motor==3.3.2

# Git & Parsing
GitPython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Notifications
slack-sdk==3.26.1
python-dotenv==1.0.0
httpx==0.25.2

# Observability
prometheus-client==0.19.0
langsmith==0.0.70

# Testing / Dev
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
>>>>>>> b44a0d4ca2b1fc73d9ca1af3689b00da856d3e5a
